requirements.txt
flask==3.0.3
numpy==1.26.4
python-dotenv==1.0.1
requests==2.32.2
langchain
langchain-community
langchain-openai
chromadb